Fixed issues in JSDB that prevented it to be built with newer versions of GCC. (Tested with GCC v6.2.1 on Arch Linux x64, GCC v4.4.6 on CentOS 6.2 x64 and GCC v4.1.2 on CentOS 5.4 x32). Also added a build script that produces binaries with debugging symbols (makejsdb_debug.csh)

+++ rs/dbf_head.cpp
@@ -334,7 +334,7 @@ bool DBFHeader::CreateFile(const char * pfname)
        }
        else
        {//that's how we normally do it
-        Field.length             = min(R->FieldLength,255u);
+        Field.length             = min(R->FieldLength,(size_t) 255u);
         Field.decimalcount       = 0;
         Field.OldSwinFieldLength = 0;
        }
